split skycoord parsers/helpers into a separate module
